+++ src/WebJobs.Extensions.OpenAI.CosmosDBSearch/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added HNSW support for CosmosDB (MongoDB) Search Provider. Refer [README](../../samples/rag-cosmos-db/README.md) for more information on usage.
 
+### Changed
+- Updated nuget dependencies
+
 ## v0.2.0 - 2024/05/06
 
 - Updated Microsoft.Azure.WebJobs.Extensions.OpenAI to 0.15.0